requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
psycopg2==2.9.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
